requirements.txt
PyYAML
jsonpath
requests
pytest
allure_python_commons
allure-pytest
python-jenkins
pytest-ordering
rsa
Flask
PyMySQL
yagmail
urllib3==1.26.18
charset_normalizer==3.2.0
Faker
gevent
pytest-tmreport
colorlog
oracledb